# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: CTEM Platform API — Continuous Threat Exposure Management
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /assets:
    get:
      operationId: listAssets
      tags: [assets]
      summary: List all assets
      parameters:
        - name: type
          in: query
          schema:
            type: string
        - name: tag
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: verificationStatus
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of assets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Asset"
    post:
      operationId: createAsset
      tags: [assets]
      summary: Add a new asset
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssetInput"
      responses:
        "201":
          description: Asset created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Asset"
  /assets/{assetId}:
    get:
      operationId: getAsset
      tags: [assets]
      summary: Get asset by ID
      parameters:
        - name: assetId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Asset details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Asset"
    patch:
      operationId: updateAsset
      tags: [assets]
      summary: Update asset
      parameters:
        - name: assetId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssetUpdate"
      responses:
        "200":
          description: Updated asset
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Asset"
    delete:
      operationId: deleteAsset
      tags: [assets]
      summary: Delete asset
      parameters:
        - name: assetId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Asset deleted
  /assets/{assetId}/verify:
    post:
      operationId: verifyAsset
      tags: [assets]
      summary: Initiate asset ownership verification
      parameters:
        - name: assetId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerificationInput"
      responses:
        "200":
          description: Verification initiated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VerificationChallenge"
  /assets/{assetId}/verify/check:
    post:
      operationId: checkAssetVerification
      tags: [assets]
      summary: Check if asset ownership is verified
      parameters:
        - name: assetId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Verification result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VerificationResult"

  # ──────────────────────────────────────────────────────────
  # ASSET GROUPS
  # ──────────────────────────────────────────────────────────
  /asset-groups:
    get:
      operationId: listAssetGroups
      tags: [asset-groups]
      summary: List asset groups
      responses:
        "200":
          description: List of asset groups
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AssetGroup"
    post:
      operationId: createAssetGroup
      tags: [asset-groups]
      summary: Create asset group
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssetGroupInput"
      responses:
        "201":
          description: Asset group created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AssetGroup"
  /asset-groups/{groupId}:
    get:
      operationId: getAssetGroup
      tags: [asset-groups]
      summary: Get asset group
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Asset group details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AssetGroup"
    patch:
      operationId: updateAssetGroup
      tags: [asset-groups]
      summary: Update asset group
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssetGroupUpdate"
      responses:
        "200":
          description: Updated asset group
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AssetGroup"
    delete:
      operationId: deleteAssetGroup
      tags: [asset-groups]
      summary: Delete asset group
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ──────────────────────────────────────────────────────────
  # SCANS
  # ──────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Asset:
      type: object
      required: [id, tenantId, name, type, value, verificationStatus, riskLevel, createdAt]
      properties:
        id:
          type: integer
        tenantId:
          type: integer
        name:
          type: string
        type:
          type: string
        value:
          type: string
        verificationStatus:
          type: string
        riskLevel:
          type: string
        tags:
          type: array
          items:
            type: string
        description:
          type: ["string", "null"]
        ipAddress:
          type: ["string", "null"]
        port:
          type: ["integer", "null"]
        isActive:
          type: boolean
        lastScannedAt:
          type: ["string", "null"]
        createdAt:
          type: string
    AssetInput:
      type: object
      required: [name, type, value]
      properties:
        name:
          type: string
        type:
          type: string
        value:
          type: string
        description:
          type: string
        tags:
          type: array
          items:
            type: string
        ipAddress:
          type: string
        port:
          type: integer
    AssetUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        tags:
          type: array
          items:
            type: string
        isActive:
          type: boolean
        riskLevel:
          type: string
    VerificationInput:
      type: object
      required: [method]
      properties:
        method:
          type: string
    VerificationChallenge:
      type: object
      required: [method, challenge, instructions]
      properties:
        method:
          type: string
        challenge:
          type: string
        instructions:
          type: string
    VerificationResult:
      type: object
      required: [verified, message]
      properties:
        verified:
          type: boolean
        message:
          type: string

    # ── Asset Group ───────────────────────────────────────
    AssetGroup:
      type: object
      required: [id, tenantId, name, createdAt]
      properties:
        id:
          type: integer
        tenantId:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        assetCount:
          type: integer
        createdAt:
          type: string
    AssetGroupInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
        description:
          type: string
        assetIds:
          type: array
          items:
            type: integer
    AssetGroupUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string

    # ── Scan ──────────────────────────────────────────────
...